Move duplicated functions to parent class DerivativeBase
