+++ Source/Grids/FourthOrderDerivatives.hpp
@@ -8,49 +8,19 @@
 
 #include "DimensionDefinitions.hpp"
 
+#include "DerivativeBase.hpp"
 #include "StateVariables.hpp"
 #include "Tensor.hpp"
 #include <AMReX_REAL.H>
 #include <array>
-#include "AMReX_Array.H"
 
 using namespace amrex::literals;
 
-class FourthOrderDerivatives
+class FourthOrderDerivatives : protected DerivativeBase
 {
-  private:
-    amrex::Real m_dx;
-    amrex::Real m_one_over_dx;
-    amrex::Real m_one_over_dx2;
-
-    [[nodiscard]] AMREX_GPU_DEVICE AMREX_FORCE_INLINE static const amrex::Real *
-    get_var_ptr(const int ivar, const amrex::Real *state_ptr_xyz,
-                const amrex::GpuArray<int, AMREX_SPACEDIM + 1> strides) noexcept
-    {
-        return state_ptr_xyz + ivar * strides[3];
-    }
-
-    [[nodiscard]] AMREX_GPU_DEVICE AMREX_FORCE_INLINE
-        amrex::GpuArray<int, AMREX_SPACEDIM + 1> static get_strides(
-            const amrex::Array4<const amrex::Real> &state) noexcept
-    {
-
-        int j_stride = static_cast<int>(state.stride.a[0]);
-        int k_stride = static_cast<int>(state.stride.a[1]);
-        int n_stride = static_cast<int>(state.stride.a[2]);
-
-        amrex::GpuArray<int, AMREX_SPACEDIM + 1> strides{1, j_stride, k_stride,
-                                                         n_stride};
-
-        return strides;
-    }
-
   public:
     AMREX_GPU_HOST_DEVICE
-    FourthOrderDerivatives(double dx)
-        : m_dx(dx), m_one_over_dx(1 / dx), m_one_over_dx2(1 / (dx * dx))
-    {
-    }
+    FourthOrderDerivatives(double dx) : DerivativeBase(dx) {}
 
     // NOLINTBEGIN(bugprone-easily-swappable-parameters)
