+++ Source/Grids/Make.package
@@ -3,6 +3,7 @@ GRTECLYN_CEXE_sources += BoundaryConditions.cpp
 GRTECLYN_CEXE_headers += BCParity.hpp \
                          BoundaryConditions.hpp \
                          Coordinates.hpp \
+                         DerivativeBase.hpp \
                          DimensionDefinitions.hpp \
                          FourthOrderDerivatives.hpp \
                          NullBCFill.hpp \
